+++ Assets/Scripts/Scene04/Result.cs
@@ -22,19 +22,19 @@ public class Result : MonoBehaviour
 
         /*
         InternalPressure = Math.Pow(1.1, (Data.CokeLevel - 1)) * Data.ShakeCount;
-        
+
         ElectricPowerGeneration = InternalPressure * ((double)(1 +((Data.WaterWheelLevel)- 1)*3) / 100) * Math.Pow(5.0, 1 +(((Data.DynamoLevel)- 1)- 1))*10.0;
-        
+
         ElectricPowerGeneration = (InternalPressure * Math.Pow(5.0, Data.DynamoLevel - 1) + Math.Pow(1.1, Data.WaterWheelLevel - 1))/10;
-        
+
         ElectricPowerGeneration = (InternalPressure + Math.Pow(1.5, Data.WaterWheelLevel - 1)) * Math.Pow(5.0, Data.DynamoLevel - 1)/10;
         */
 
         ElectricPowerGeneration =  InternalPressure*BigInteger.Pow(Data.WaterWheelLevel,3)*BigInteger.Pow(Data.DynamoLevel,100);
 
         CurrentElectricPowerGeneration = ElectricPowerGeneration;
-        CurrentElectricPowerGenerationsmall = (double)ElectricPowerGeneration;       
-        Money = ElectricPowerGeneration;             
+        CurrentElectricPowerGenerationsmall = (double)ElectricPowerGeneration;
+        Money = ElectricPowerGeneration;
         MoneyText.text = string.Format("獲得金額:{0}円", CurrentMoney);
         ElectricPowerGenerationText.text = string.Format("発電量:{0}kw", CurrentElectricPowerGeneration);
         ShakeCountText.text = string.Format("振った回数:{0}回", Data.ShakeCount);
@@ -42,6 +42,8 @@ public class Result : MonoBehaviour
         SecondText.text = string.Format("振った秒数:{0:0.00}秒",Data.ShakeTime);
 
         Data.money += Money;
+
+        SaveLoad.Save();
 
     }
 
@@ -60,12 +62,12 @@ public class Result : MonoBehaviour
             {
                 CurrentElectricPowerGenerationsmall -= (double)ElectricPowerGeneration * (Time.deltaTime / 2.0f);
             }
-            
+
             if (CurrentElectricPowerGenerationsmall <= 0)
             {
                 CurrentElectricPowerGenerationsmall = (double)0;
             }
-            
+
             ElectricPowerGenerationText.text = string.Format("発電量:{0:0.00}kw", CurrentElectricPowerGenerationsmall);
 
             }
@@ -76,33 +78,33 @@ public class Result : MonoBehaviour
             {
                  CurrentMoneysmall += (double)Money * (Time.deltaTime / 2.0f);
             }
-            
+
             if (CurrentMoneysmall >= (double)Money)
             {
                 CurrentMoneysmall = (double)Money;
             }
 
-            MoneyText.text = string.Format("獲得金額:{0:0}円", CurrentMoneysmall);          
-            
+            MoneyText.text = string.Format("獲得金額:{0:0}円", CurrentMoneysmall);
+
             }
 
-            
+
             else{
             if (CurrentElectricPowerGeneration > 0)
             {
                 CurrentElectricPowerGeneration -= ElectricPowerGeneration / (BigInteger)(2.0f / Time.deltaTime);
             }
-            
+
             if (CurrentElectricPowerGeneration <= 0)
             {
                 CurrentElectricPowerGeneration = 0;
             }
-            
+
             if (CurrentMoney < Money)
             {
                 CurrentMoney += Money / (BigInteger)(2.0f / Time.deltaTime);
             }
-            
+
             if (CurrentMoney >= Money)
             {
                 CurrentMoney = Money;
@@ -115,7 +117,7 @@ public class Result : MonoBehaviour
             }
         }
 
-        /*DOTween.To 
+        /*DOTween.To
             (
                 () => CurrentElectricPowerGeneration,       //何に
                 (x) => CurrentElectricPowerGeneration = x,  //何を
@@ -123,7 +125,7 @@ public class Result : MonoBehaviour
                 1                                        //どれくらいの時間
             ).SetEase(Ease.Linear);
 
-            DOTween.To 
+            DOTween.To
             (
                 () => CurrentMoney,         //何に
                 (x) => CurrentMoney = x,    //何を
