+++ Assets/Scripts/Scene1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 64c11e8f758591b4eadfc054ae01f4be
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
