+++ Assets/Scripts/Scene1/Save.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Save : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+        SaveLoad.Load();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+
+    }
+}
