+++ Assets/Scripts/Scene1/Save.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 142223e406d050140b520204bb5e239c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
